Criado o cadastro e Login de Usuário

+++ src/main/java/scc/vigilancia/comunitaria/models/Community.java
@@ -2,8 +2,6 @@ package scc.vigilancia.comunitaria.models;
 
 
 import lombok.Data;
-import lombok.Getter;
-import lombok.Setter;
 
 import javax.persistence.*;
 import java.util.List;
@@ -15,7 +13,7 @@ public class Community {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(name = "nome")
     private String name;
